Organização da estrutura do código em diferentes arquivos e bibliotecas
