+++ include/bme280.h
@@ -1,0 +1,19 @@
+#ifndef _BME280_H_
+
+#define _BME280_H_
+
+#include "main.h"
+
+
+#include <Adafruit_BME280.h>//Sernsor BME280
+#include <Adafruit_GFX.h>
+
+//Variable of Sensor (Temperature, pressure and Humidity)
+
+#define SEALEVELPRESSURE_HPA (1013.25)
+
+void BME280_Sleep(int device_address);
+void BME280_read_mean(int number_samples_mean, float *temperature_rtc, float *pressure_rtc, float *altitude_rtc, float *humidity_rtc);
+void BME280_status();
+
+#endif //_BME280_H_
